[527603] Create always library of forte and allow many types of build
Forte is always compiled as an internal library and then the executable
adds the main to the compilation. Static and shared libraries can also
be built altogether. Tests run against the same library

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=527603

+++ tests/core/fbtests/fbtester.cpp
@@ -297,7 +297,7 @@ bool CFBTester::createDataInputConnections(){
   else{
     printf(
         "*** FB test preparation: Wrong amount of input data given! Expected: %d, got: %lu\n",
-        interfaceSpec->m_nNumDIs, mInputDataBuffers.size());
+        interfaceSpec->m_nNumDIs, static_cast<long unsigned int>(mInputDataBuffers.size()));
     retVal = false;
   }
 
@@ -331,7 +331,7 @@ bool CFBTester::createDataOutputConnections(){
   }
   else{
     printf("*** FB test preparation: Wrong amount of output data given! Expected: %d, got: %lu\n",
-        interfaceSpec->m_nNumDOs, mOutputDataBuffers.size());
+        interfaceSpec->m_nNumDOs, static_cast<long unsigned int>(mOutputDataBuffers.size()));
     retVal = false;
   }
